add file check

+++ GP Mobile/Android/GPVision/src/com/gpvision/fragment/ChooseFileFragment.java
@@ -21,7 +21,7 @@ public class ChooseFileFragment extends BaseFragment {
 
 	public static final String TAG = ChooseFileFragment.class.getName();
 	private static final int REQUEST_CODE_CHOOSE_FILE = 2030;
-	private static final int MAX_FILE_SIZE = 1024 * 1024 * 1024;
+	private static final long MAX_FILE_SIZE = 1024 * 1024 * 1024L;
 	private File file;
 	private OnChoseListener listener;
 	private TextView fileInfo;
